+++ nodeconductor/structure/tests/unittests/test_signals.py
@@ -50,8 +50,7 @@ class ServiceProjectLinkSignalsTest(TestCase):
 
     def create_service(self, customer, available_for_all):
         service_settings = factories.ServiceSettingsFactory(shared=False)
-        return test_models.TestService.objects.create(name='test',
-                                                      customer=customer,
+        return test_models.TestService.objects.create(customer=customer,
                                                       settings=service_settings,
                                                       available_for_all=available_for_all)