+++ CHANGELOG.rst
@@ -1,8 +1,12 @@
 CHANGELOG
 =========
 
-Release NEXT
-------------
+Release 0.125.0
+---------------
+- Change external and internal ips to property. Force all models inherited from VirtualMachineMixin to define external
+and internal ips properties.
+- Remove name from service, service settings name is used instead.
+- Introduce and connect certification model to service settings.
 
 Release 0.124.0
 ---------------